+++ WheaterApp/app/src/main/AndroidManifest.xml
@@ -2,8 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.weatherapp">
 
-    <uses-permission android:name="android.permission.INTERNET"/>
-
     <application
         android:name=".SampleApplication"
         android:allowBackup="true"
