+++ WheaterApp/app/src/main/java/com/example/weatherapp/viewmodels/CityMainViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.domain.entities.City
-import com.example.domain.usecases.GetCityByNameUseCase
+import com.example.domain.usecases.GetCityByIdUseCase
 import com.example.domain.usecases.JSONDataUseCase
 import com.example.weatherapp.contracts.CityContract
 import com.example.weatherapp.utils.Data
@@ -13,7 +13,7 @@ import com.example.weatherapp.utils.Status.DONE
 import com.example.weatherapp.utils.Status.INIT
 
 class CityMainViewModel(
-        private val getCityByNameUseCase: GetCityByNameUseCase,
+        private val getCityByIdUseCase: GetCityByIdUseCase,
         private val JSONDataUseCase: JSONDataUseCase
 ) : ViewModel(), CityContract.ViewModel {
 
@@ -32,7 +32,7 @@ class CityMainViewModel(
         mutableMainState.value = Event(Data(status = INIT, listOfCities = JSONDataUseCase.getJSONData()))
     }
 
-    override fun getCityId(name: String): Int = getCityByNameUseCase(listOfCity, name)
+    override fun getCityId(name: String): Int = getCityByIdUseCase(listOfCity, name)
 
     companion object {
         var listOfCity = mutableListOf<City>()
