+++ WheaterApp/data/build.gradle
@@ -28,9 +28,4 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(path: ':domain')
-
-    def lib = rootProject.ext.generalDependencies
-
-    implementation lib.retrofit
-    implementation lib.retrofitGson
 }
