--- WheaterApp/data/src/main/java/com/example/data/utils/Utils.kt
@@ -1,5 +1,0 @@
-package com.example.data.utils
-
-const val NOT_FOUND = "NOT FOUND"
-const val DEFAULT_ID = 0
-const val DEFAULT_TEMP = 0.0
