+++ WheaterApp/domain/src/main/java/com/example/domain/entities/City.kt
@@ -1,9 +1,5 @@
 package com.example.domain.entities
 
-import com.example.domain.utils.DEFAULT_ID
-import com.example.domain.utils.DEFAULT_TEMP
-import com.example.domain.utils.NOT_FOUND
-
 data class City(
     val id: Int = DEFAULT_ID,
     val name: String = NOT_FOUND,
@@ -11,4 +7,10 @@ data class City(
     val main: String = NOT_FOUND,
     val description: String = NOT_FOUND,
     val temp: Double = DEFAULT_TEMP
-)
+) {
+    companion object {
+        const val NOT_FOUND = "NOT FOUND"
+        const val DEFAULT_ID = 0
+        const val DEFAULT_TEMP = 0.0
+    }
+}
