--- WheaterApp/domain/src/main/java/com/example/domain/repositories/CityJSONRepository.kt
@@ -1,10 +1,0 @@
-package com.example.domain.repositories
-
-interface CityJSONRepository {
-
-    fun isNotEmptyData(): Boolean
-
-    fun getJSONData(): MutableList<String>
-
-    fun setJSONData(listOfString: MutableList<String>)
-}
