--- WheaterApp/domain/src/main/java/com/example/domain/servicies/WeatherCityService.kt
@@ -1,6 +1,0 @@
-package com.example.domain.servicies
-
-interface WeatherCityService {
-
-    fun getCityById(id: Int): Boolean
-}
