+++ WheaterApp/domain/src/main/java/com/example/domain/usecases/GetCityByIdUseCase.kt
@@ -1,6 +1,8 @@
 package com.example.domain.usecases
 
+import com.example.domain.entities.City
+
 interface GetCityByIdUseCase {
 
-    operator fun invoke(id: Int): Boolean
+    operator fun invoke(listOfCity:MutableList<City>, name: String): Int
 }
