--- WheaterApp/domain/src/main/java/com/example/domain/usecases/GetCityByNameUseCase.kt
@@ -1,8 +1,0 @@
-package com.example.domain.usecases
-
-import com.example.domain.entities.City
-
-interface GetCityByNameUseCase {
-
-    operator fun invoke(listOfCity:MutableList<City>, name: String): Int
-}
