+++ WheaterApp/domain/src/main/java/com/example/domain/usecases/implementation/GetCityByIdUseCaseImpl.kt
@@ -1,13 +1,17 @@
 package com.example.domain.usecases.implementation
 
-import com.example.domain.servicies.WeatherCityService
+import com.example.domain.utils.ZERO
+import com.example.domain.entities.City
 import com.example.domain.usecases.GetCityByIdUseCase
-import org.koin.core.KoinComponent
-import org.koin.core.inject
 
-class GetCityByIdUseCaseImpl() : GetCityByIdUseCase, KoinComponent {
+class GetCityByIdUseCaseImpl: GetCityByIdUseCase {
 
-    private val weatherCityService: WeatherCityService by inject()
-
-    override fun invoke(id: Int) = weatherCityService.getCityById(id)
+    override fun invoke(listOfCity: MutableList<City>, name: String): Int {
+        for (i in ZERO until listOfCity.size) {
+            if (listOfCity[i].name.equals(name)) {
+                return listOfCity[i].id
+            }
+        }
+        return ZERO
+    }
 }
