+++ WheaterApp/domain/src/main/java/com/example/domain/utils/Utils.kt
@@ -5,6 +5,3 @@ const val ID = "id"
 const val NAME = "name"
 const val COUNTRY = "country"
 const val COUNTRY_AR = "AR"
-const val NOT_FOUND = "NOT FOUND"
-const val DEFAULT_ID = 0
-const val DEFAULT_TEMP = 0.0
